Save country ID to orders
When an order is loaded from the database after returning from an
off-site payment processor, it needs the country ID to load the
country information.  This adds the country ID to each address in
the order table.

Addresses bugs where country information is missing after returning
from a payment processor.