New HTML classes
These are the new HTML classes to replace the functionality in the
includes/functions/html_output.php files.  Most of these relate to
form inputs, as that is primarily what the html_output.php files
provide.  Also replaces the catalog href_link.php and image.php
files.

Unifies the catalog and admin functionality.