Chain hooks
This adds the capability to pass input into a set of hooks such that
the results of each is passed to the next in the "chain".  This is
needed for the Href class.

Also removes the concept of aliases.  Now hooks are called by action
regardless of group.  This simplifies the handling of pipeline and
siteWide hooks.

The generate method is modified immediately to no longer take a
group argument.

The call method has been deprecated in favor of a cat method that
does not take a group argument.  For compatibility reasons, call
simply calls cat at the moment.  When all the call references are
replaced in core, it will get a deprecation trigger.

Refactors registration and loading to accomodate the removal of the
group argument.

Adds a new `set` method to allow code to register individual hooks.