Deprecate hook aliases
Previously pipeline hooks used the concept of aliases, registering
the pipeline group as the page group instead.  The hooks class was
modified to no longer do this.  This alters the things that call
generate to no longer pass the group argument.

This causes a conflict between the checkout pipeline and the system
hooks in application_top.php.  That conflict is resolved by renaming
the checkout pipeline's action to differentiate it from the system
hooks.