Improve handling of repeated inputs
If the page is submitted and the same menu is shown again, we will
usually want to display the value from the first submission.

This changes the class to default to doing that but still allows for
developers to override this (by clearing the selection or explicitly
setting it to something else).  Also allows to easily set a default
value that will be overridden by the submitted value if it exists.

Adjusts the compatibility functions to reflect these changes.