Replace tep_get_all_languages
The language::load_languages method almost does the same thing as
tep_get_all_languages, but it puts the code in the key rather than
the data.

This leaves the code in the data, although it also leaves it in the
key.  It also renames the method to language::load_all, which is
less redundant.

Some minor cleanup and simplification.