Remove PayPal notice
PayPal duplicates the currencies::format_raw method.  A bug in that
method was fixed but not copied over.  This changes formatCurrencyRaw
to use the currencies method.

Long term, a better plan would be to eliminate uses of the PayPal
version in favor of the currencies method directly.  But this at
least gets us past the immediate problem.