New password classes
To replace the password function files, migrating to classes.

The new classes use the modern password_hash and password_verify
functions.  When a customer with an older password logs into the
site, the password will be updated.  These functions also support
more configuration.

The Password class is the actual one used for authentication.  The
old_password class is just for handling legacy customers.  Some
customers may also be using the PHPass hashing, which was already
a class.  The apache_password class is only used for htpasswd files.