Fall back to fsockopen if curl and allow_url_fopen off
The simplest from our perspective is if the allow_url_fopen
configuration is set.  Then we can just load remote files as if
they were local.

If not, we fall back to using curl.  This adds another fallback if
curl is not available.  This second fallback uses fsockopen
manually to fetch the URL.

Also refactors to make a separate loading function for those times
when we aren't loading XML.  And if allow_url_fopen is off, we use
that method from the load_xml method.  Because the load_xml method
falls back to parsing a string fetched via some other manner.  The
duplicate manners are available in the load method.