Modularize admin login actions
Traditionally admin has processed actions with a switch statement.
So to add an action requires editing the code of the switch.

This changes to allow each action to be its own file.  So adding a
new action is as simple as adding a new file with the correct name.

This also puts the boilerplate code for action processing into its
own file.  And it adds a new verification step for any action it
processes.  This is intended to prevent Cross Site Request Forgeries.

Also introduces a new button method to the Admin class.  This is
intended to allow more control of admin links for add-ons that
restrict admin access partially.