Default image to be responsive
The functions would default responsive to true and then set it.
The new class required it to be set manually.

This changes the default for the class to be responsive, while
still allowing it to be turned off if necessary.  This allows
direct uses of the class to have the same default behavior as the
functions that it replaces.