Improve default images
When the default image is turned on, the code checks if the image
file exists.  But if the image src is an absolute path, this breaks.

This change creates two separate kinds of image prefixes: file
system and web.  The file system or path prefix is just the location
in the file system.  The web prefix is a URL.

So now, we can check the file system path for the file and output
the web path in the HTML.