Update for new autoloader
The translation portion of the new autoloader requires additional
handling.  In particular, we want the Text class to be loaded
before the template object is built.  Because otherwise we can get
in an endless loop, as the autoloader tries to call itself.

Same problem with the template object creation and setting the
translator.  We need it built before the translator is set so as
to avoid a situation where the translator triggers the building of
the template object, which invokes the autoloader and translator...

Also switches to call the Request class for the session checks.
This is preparation for using hooks to implement configuration.
Also, two of the session checks already call into the Request class.