Move tax language constant out of english.php
Since Tax can be used either in catalog or admin, we either need to
duplicate the language constant or move it out of the catalog-specific
file.  This moves it out so that both catalog and admin will share
the same translation.

This should be about the same difficult for translators now and easier
to maintain in the future.