Compatibility functions for general.php
This modifies the includes/functions/general.php files in both admin
and catalog to call the new class method replacements.  These will
be left through 1.0.9.0 for compatiblity.

As the core callers are modified to no longer call these, they will
be marked as deprecated.  Those that are no longer used are marked
now.  The others will be migrated as practical.