Use quicker test if available
We can use a regular expression to check for a variety of locale
variations.  But the simplest check, which will often work, is to
see if it matches the two letter language code.  This saves us an
entire loop when it works and is a quick check when it doesn't.