Add missing function
The old way was to pass a string like 'class="form-control"' to the
function.  The new way is to call append_css or set a class value
in the parameters array.

We have to convert from the old way to the new in multiple places.
This just creates a function to do that for us and migrates the
existing uses to it.

Like the other phoenix_ functions, this will be deprecated in the
future when it is no longer needed.