Remove obsolete function files from install
We no longer load these files, so there is no reason to keep them.

The functions within them have been replaced with autoloaded class
methods.  Most are from the regular system classes, but a few
install-specific functions are in the installer itself.

This finishes eliminating the old naming scheme from the installer.