Eliminate tep_array_to_string
In the new way, we pass the query parameters as an array rather than
a string.  So we no longer need a function to convert an array to a
string.  We do that in the Href class itself.

This changes to use the new Href class and makes other changes to
modernize the affected files.