Remove obsolete class file
If we need the PasswordHash class, we will load the catalog version.
We no longer need a separate admin copy in the core install.

Note that we only use the PasswordHash class to verify legacy
passwords before replacing them with hashes created by the newer
PHP password_hash function.  So new installs should not ever be
using this class.  Only older installs.  And not all of the class
will be used.