Modernize split page results in a backward compatible way
Migrates function calls to new methods.

Modifies the display_links method to take a link as an argument.
If no link is provided, it uses a default (which will work for all
core calls).  If called with a string as previously, it will parse
the string for the parameters.

In the future, support for calling with a string can be removed.