Fix set as primary display in Chrome
Set to have a blank CSS class so as to not be form-control by default.
This makes the checkbox display in Chrome.

Not entirely clear why it was not working.  But this makes the HTML
more like it used to be.

Also modernizes the echo statements and switches from function to new
class method.  This is forward compatible.