Add SQL to error log
When a SQL query errors, we log the error number and message.  This
also logs the query that errored.  Because the error message is
sometimes clearer when viewed in the context of the actual query
that generated it.

This was specifically added because of an error that arose when two
separate queries were jammed into one request.  Easy to diagnose
with the actual SQL.  But the error message excerpted a perfectly
legitimate section of the second query.