Fix GDPR module compilation
Because public_title is a property of the abstract base class,
overriding it sends the compiler down a path where it attempts to
use the abstract class' value for CONFIG_KEY_BASE, which is invalid.

This overrides the value in the constructor instead, which does not
trigger the same problem.