Add concept of infobox actions
Previously we added the concept of action segments that could be
automatically loaded.  This replaces the giant switch statement at
the beginning of most admin page files.  So adding a new action is
as simple as adding a new file.

This change does the same thing for the infobox action switch at
the end of each file.

Refactors the locate_action method for code reuse purposes.

Picks up two minor tweaks to the code that are unrelated.