Use a new infobox component rather than the old box class
To support the change from a switch statement to a collection of
files, this switches from the old box class to a new infobox
component.  This will be simpler than the original plan to migrate
the switch into a box function attached to the table definition.
Also, this allows for new actions to be added as simply as creating
a file (or two if the action causes behavior both at the top and
bottom of the file).

This is also part of a general move away from using classes for
display.  This is more like a MVC application, with the display in
separate files from the data and processing.

In the long term, this should allow the box and tableBlock classes
to be deprecated.  Which is helpful because the two reinforce the
table-based layout.  With this version, changing that layout would
be as simple as replacing the component file.

The component is also simpler than the tableBlock version, which
is over-engineered with capabilities that are never used.  And of
course it only uses one file where the other uses box, tableBlock,
and a template file.