Add convenience method to get zone name
We want to be able to take either a numeric zone ID or a zone name.
The Zone::fetch_name requires parameters to handle that.  This
maps what we would get to what the method needs.

This restores the old behavior:  if the zone ID matches a valid
zone, show the zone name.  Otherwise, treat the zone ID as if it
were the zone name.

It also goes further and posts a warning message if the zone does
not match the country.