Link to catalog directory, not the current page
By default, we link to the current page if no page is set.  But
when switching to catalog, we want to link to the catalog directory.
Otherwise we attempt to link to an admin page in the catalog
directory, which doesn't work with most pages.

This changes from not setting the page at all to setting it to a
blank string.  The blank string keeps it from defaulting, so that
we get directed to the catalog directory.