Better special product selection
We only want to prevent people from seeing the new specials page.
If they have no undiscounted products, they may still want to view
discounted products under other actions.

Added action processing for the new case to handle that.

Modified the select menu creation to get the list of discountable
products if that had not yet been done.

Named $discountables to be more descriptive.

Remove the code to mask out products on special.  We do that in the
database query directly now.  So we don't need to do a second query
to do the same thing.  That code was orphaned and its result never
used.