Add web_id to table_definition
When we make the page links, we want to include all the current
parameters except the page (which we are changing) and the row ID
(if set).  But to exclude the row ID in general, we need to know
what it is.  Each page uses a different one.

This adds parameters for the web version of the row ID key and the
database version.  Those are usually different.

Updates the specials page now.  Other pages that weren't modified
in this release will be  modified in a future release.

Also adds a default procesing function for rows.  With the web and
database names for the ID, we can write a general version that will
usually work.  We still allow table definitions to specify their
own for those times when the default isn't quite right.