Fix low order fee module install
The use function has special handling for the -> case.  It creates
an object with the same name as the left hand side and calls the
method with the same name as the right hand side.  This is weird
but works.  Unlike the set function, this can't be replaced with a
Guarantor::ensure_global call.

This change reverts back to the working form.