Update navigation menu
We moved from a model of certifying add-ons to just certifying the
developers.  So change the language text to make this clearer.
Added a new entry, since we now have a separate add-ons area.

Some people were not realizing that they could go to the forum for
help, so making that clearer so as to reduce contacts.

Overall, this is one more link but less, hopefully more readable
text.

Minor tweaks to comments and formatting for readability.