Modernize
We have rebranded as CE Phoenix Cart, so updating the language file
to match.

Updating from define to const.  A side effect of this is that some
code that was in the language file moves into the template file.
That code is also updated to use the new object method rather than
the legacy functions.