Fix non-form actions on product_attributes page
The new form processing checks the formid to see if it matches the
sessiontoken.  So of course, we have to make sure the formid is in
the request.  If we're using a form, this happens automatically.
If we're using a link, this needs to be done manually.  This creates
the link ahead of time and makes use of it where appropriate.