Allow select_multiple to select none
Sometimes when people are using a multi-select, they want to be able
to clear out the existing entries.  The original version doesn't
support that, as when nothing is selected, nothing is sent.  It's as
if the checkboxes don't exist.

This adds a hidden input with an empty string.  If any of the
checkboxes are ticked, they will send an array instead.  But if none
of them are, it will fall back to sending an empty string.

The select_at_least_one method retains the original functionality.
If anyone needs it, they can switch to that method.