Modernize admin currencies page
The language file for the admin currencies page had an embedded link
in it.  The old way to handle this was to simply embed the call to
the link function in the value by string concatenation.  The new way
is to use string formatting placeholders, since the link will be the
same for all languages.

Updated the page to use action files and a paginated table.

General modernization and clean up.