Fix create_account when field invalid
When a field is invalid in create_account, we want to continue and
process the remainder of the fields.  This allows everything to be
set in the customer_details array.  So the user will see the same
data as was entered without having to redo everything.

The issue here was that it would short circuit once successful was
false.  So the remainder wouldn't get processed.  Changing the
order fixes that.