Hook by originally requested action
Sometimes, we change from one action to another inside the action
handler.  When this happens, we probably want to hook by the
original action rather than the new one.  If not, we should have
mapped with an earlier hook (e.g. the preAction hook listener).

This modifies the action processor to save the original action
before sending to the action handler.