Modernize admin/reviews
This should also resolve

https://github.com/CE-PhoenixCart/PhoenixCart/issues/16

Not so much the reported issue (that the admin can insert Javascript
into the text field that will display both in admin and in catalog).
But the title issue of allowing XSS (via CSRF) on the admin reviews
page.  This uses process_action to check that the request was made
with knowledge of the session token injected into the form.