Slim down account creation
We inherited a number of fields on account creation that seem to be
intended more for demo purposes than for a typical shop.  I.e. they
exhibit all behaviors that someone might want to have, not the
typical behaviors that most stores would want to have.

This removes telephone, newsletter, and password confirmation from
the default registration process.  If stores want these, they can
install the relevant modules themselves.  For existing stores, no
action is necessary.  Presumably they've already installed or
uninstalled modules as appropriate.

We may later want to remove the address from the default as well.
But currently, the store doesn't function well if accounts do not
have at least one address.  And of course, shipping requires an
address.

This reduces the number of clicks to register in the default install,
which should make testing easier.