Only use Helper option when Enforcing Selection
In the options_attributes modules for PI layout and PI content, there
is an option to show Helper text in the select menu.  If that is
selected and Enforce Selection is not, then adding products to the
cart breaks.

This changes to ignore the Helper text option when Enforce Selection
is off.  This keeps things in a working state, regardless of what the
store owner chooses.

In the future, we may rationalize the module options.  E.g. by
combining the Helper and Enforce Selection options into a single
option with only the three valid states.  That would better align
the capabilities of the code with the possible options.

A more complicated solution would make the Helper menu item return
the same ID as the first attribute when Enforce Selection is off.
That would give the same add-to-cart behavior for both settings
while allowing the store owner to pick all four states with
different behavior for each.  Different in terms of what the
interface would display at least.

This was picked as the current solution as being the lowest impact
change.  It simply ignores selection of an option combination that
we would recommend against using anyway.