Maintain shipping selection after editing
If someone makes a shipping selection, proceeds to the confirmation
page, and then goes back to the shipping selection page, the current
behavior is to force the customer to manually select a shipping
method again.

This changes the behavior to maintain the previous selection.  The
customer may change it, but by default, it is assumed that they want
to keep the same one.

This still maintains the property of not making a shipping selection
for the customer on the first visit to the page.

Note that this also means that we don't select the cheapest method
by default.  That was previously done to avoid having to implement a
force selection in Javascript.  However, we now have a force
selection from Bootstrap.  So that is no longer necessary.