Migrate MATC to a customer data module
The MATC is implemented as a hook.  This made sense when it was
originally created, as customer data modules did not exist then and
it was a separate module.  However, it is currently part of the
default install, as it is require in GDPR jurisdictions.  But it is
not required in non-GDPR jurisdictions.

Migrating it to a customer data module makes it possible to turn it
off and increases the configuration options.