Autoload languages files for hooks in templates
A while ago, we started autoloading the language files for class
and hook files located under the includes directory.

Also a while ago, we made it possible to load hook files for specific
templates.  For example, most of the core hooks moved to default.

What we had not previously done was made it autoload language files
for hook files in the templates directory.  But this change now does
that.