Only trigger warnings on database error messages
Triggering an error halts execution.  This can happen prior to the
message logging completely.  Changing this to a warning ensures that
the message can log completely.  Then we die manually after logging
the message.

This shouldn't matter in normal usage.  But when there is a database
error, this version should produce better logs.