Simplify array query parameter retention
Rather than allow array query parameters, this moves the flattening
to the retention step.  This makes the retention step slightly more
complicated, but it makes the query string building simpler.

This also makes it so that URL rewriters that use the query string
parameters from the Href will get the flattening without additional
changes.  And they can continue to assume that query parameters
will be comprised of key/value string pairs.