Modernize admin/backup
This does most of the usual:  move action specific logic into
admin/includes/actions/; replace functions with methods; update
the code to the latest standard; convert jQuery to vanilla JS.
It does not make the backup list paginated.  It does change the
data manipulating actions to use POST forms instead of GET links.

To simplify the code, this also migrates from manually parsing
the SQL into statements and calling query on each to calling
mysqli->multi_query on the whole file.  That is now done in a
new class named sql_file.