Better download handling
This promotes all warnings, notices, etc. to fatal errors, so they
don't appear in the downloaded file.

This suppresses content from whitespace, etc. from appearing in the
downloaded file.

This moves the DOWNLOAD_BY_REDIRECT code into its own file, so if
not used, it doesn't need to be parsed.

This removes the umask, as that affects all processes on the server.
If needed, people can implement this via hook.  But it's out of core.

Switches to scandir to simplify the cleanup portion.  Note that we
trigger a clean up on every download, so there should only be one
such directory in the typical case.  Perhaps race conditions might
lead to more than one (e.g. if a file couldn't be deleted because it
was open for downloading at the time).

Simplify the random name generation.  We don't need to maintain an
extra index variable, as PHP strings already track their length.

This also modifies the content-disposition to support file names that
include spaces by quoting the filename.

Note:  most store owners will find the readfile method to be more
robust than the redirect method.  Yes, it uses more server resources,
but it uses them for a purpose.  In particular, it requires that a
login session exist to download the file.  The redirect method allows
unlimited downloads of the file until a second person attempts a
download through the application.