Call catalog hooks for catalog links in admin
We have a method to create a catalog link in admin, but it still
calls the admin hooks.  This modifies it to call the catalog hooks
in admin.

The primary use for this is for URL rewriters.  This allows them to
change the catalog links that appear in admin to match how the links
appear in catalog.  This is most important for notifications and
things like sitemaps.  It is also convenient for other links, so
that administrators don't have to constantly switch to catalog to
view their actual links.