Fix language deletion
Fixes a bug where the default language check logic was backward (so
that the error would show when the language was not the default).

Converts deletion to a POST form rather than a GET link.  Not only
is this good practice in general (write operations should be POST
rather than GET), it makes the action processing work without further
manipulation.