Standardizing on Timor-Leste
Timor-Leste had been previously added instead of changing East Timor to
Timor-Leste.  Then standardized both on East Timor.  As Timor-Leste is
the name used in the country itself, changing to that.  It is reasonable
to guess that most people shipping there will be from there and using
the Portuguese version of the name.

This also adjusts the country codes to the current TL/TLS.

This solution creates a duplicate entry.  However, the duplicate entry
was already present, just not obvious.  Since removing countries breaks
existing stores, we should be careful about doing it.  At some time in
the future, we can work out a way to remove all the obsolete and duplicate
entries at once.

Note:  individual stores may of course remove entries now.  What I'm
saying is that we should try to avoid a situation where some stores
have the US as 223 and others have it as 234.  Ideally all stores
would use the same countries_id for the US and other countries.