Only show specials sidebox if at least one special enabled
Previously, if there were no specials enabled, it would return an
offset of 0.  Then it would query again and try to find a special.
But we already know that there is no special to find.

Instead of returning an offset of 0 when there are no specials
enabled, this returns an empty result set.  So we only send one
query rather than two.