Remove fatal after expiring action records
The old function called mysqli_affected_rows with a default parameter.  When this was updated to call mysqli_affected_rows directly, the parameter wasn't put in the call (a necessary step).  This occurs after the expiration, so someone could just reload the page.  But better to just get rid of the fatal.