Fix layout module install
For content modules, group is the directory under content where the
module files are stored.  For layout modules, group identifies both
the page and slot where the module goes.  Thus, for content modules
group and page are the same thing.  But for layout modules they are
different.  This provides a definition of page for layout modules
that works consistently.

Basically, we get the page of a layout module from its file path.
That works consistently and allows us to find other information from
the page.

This also updates the configuration directly from the cfgm_layout
rather than trying to return to cfg_modules and use the standard
logic.