Show infobox for module whenever one selected
We want to show an infobox for a module when it is selected.  If we are
looking at the list of modules to install, we show an install button.
Otherwise, show an edit button.  Note that this doesn't try to protect
against people manually editing the URL.  If they do, weird things will
happen.