Permanent PHP 8 fixes where possible
We previously put in a bunch of temporary PHP 8 fixes.  Where
possible without losing PHP 7 compatibility, this migrates to the
permanent fixes.

The biggest problem is that PHP 8 supports union types and PHP 7
does not.  So we can't fully support overriding/implementing PHP 8
methods with union return types until we drop PHP 7 support.