No warning on lack of data
We don't want to warn if there is no data for a particular session
key, only if the database query itself fails.  Return an empty string
when there is no data.  Will continue to warn on a database failure
as part of a cascade of messages, as this seems to be the intended
(by PHP) behavior.