Rewrite time logging to sanitize formatted time before writing
This renames logger to Logger, as is standard in Phoenix.

Since there is no ongoing state, there is no reason to instantiate an
object.  This changes to static functions.

Rearrange logic to not echo if display not requested.  Previously
echoed a null (which converts to an empty string).

Sanitizes the formatted time string to prevent the creation of
malicious code files by bad actors with admin access.  This resolves
https://github.com/CE-PhoenixCart/PhoenixCart/issues/43`