Change New to Latest (Products)
There is a module that displays the X most recent products.  It says
that it shows the "New Products for [month name]", but it doesn't.
Some of the products will be from previous months or not all the new
products for the current month will be shown.  This fixes that.

In addition, strftime is being deprecated and this eliminates that
usage.