Deprecate strftime
We had been using strftime to localize dates, including names, but
it has been deprecated by PHP.

This replaces it with IntlDateFormatter.  Unfortunately, the `intl`
extension is not always installed in PHP.  We provide a polyfill
(based on the one used by Symfony) for when that happens.  That
polyfill only works correctly for en.  People who want true
localization will need to use a PHP with intl installed.

This set of changes are not backwardly compatible and need to go in
one block.

People using custom translations instead of the defaults will have
to adjust their english.php files with these changes manually.

People using add-on translations may want to ask the maintainer to
update the add-on with equivalent changes.