Move product card buttons into hook
Currently, the product card component manually adds buttons if
desired by the caller (the default is no buttons).  This moves that
behavior into a hook so that it can be turned off.

This uses a per-template hook as the buttons include HTML specific
to a template.  Different templates can override that behavior.

Also includes a change to allow extra content to be overridden more
easily using the same hookpoint.