Login / Create Account changes
This change adds an extra admin setting allowing the shop owner to decide what page should be shown when a non-logged-in customer decides to checkout;  create_account (good for shops with mostly new clientelle) or login (good for shops with mostly returning customers).  Each shopowner can decide what is best for their demographic.

This also revamps the login.php page and the password_forgotten.php page to make it prettier.