Split PI Gallery Module
This splits the Gallery PI Layout module (which displays the main image and the extra images (if any)) into two separate modules.

A module for the main image and a module for the extra images.

This also introduces a Hook which creates the logic for the modal popup window showing the Products Album.

This split of modules allows more flexibility in the product page PI layout.

Note that the normal Gallery Content Module remains unchanged.