Fix Installer
This fixes the installer so that error messages are returned if the user has inserted incorrect data or if there are other problems.

The error messages are somewhat cryptic, but translating those would be a task that may cause significant issues (eg if a given error message is not given a translation).