Only change rows in the same table
The previous code changed rows in all tables on the page.  We would prefer to limit the change just to the current table so as to allow multiple selection tables on the same page.  

Also adds double quotes to make it clear that it is intentional not to use single quotes.