Fix table alias in customer statistics
We previously switched how we built the table aliases for the new product search functionality.  The customer statistics build a query partially manually so as to allow for a custom grouping and order.  This switches that to the new table alias determination rather than the old (which no longer works).