Increase minimum required version to 7.3
We have been supporting a minimum version of 7.0, but that's becoming increasingly infeasible.  All the 7.* versions are out of even long term support now.  Dropping 7.0 through 7.2 so that in the future, we can remove the changes specific to those versions (e.g. the multiple versions of the Cookie class).  

In the (not distant) future, expect us to increase the minimum to 7.4 and 8.0 for similar reasons.