Update minimum PHP version requirement to 7.3
We have already published a minimum version of PHP 7.3.  This makes the installer enforce that.  

Note that we plan to increase the requirement to 8.0.  We are doing so slowly to minimize the issues that may arise in any one version change.  

Increasing the minimum to 7.3 will allow us to simplify the Cookie class, which currently has two different versions depending on before or since 7.3.  If anyone needs to support 7.2 or earlier, they could override with the older version and hack the installer.  However, this allows us to simplify core going forward.  

Note that PHP no longer supports any 7.x version.  People are strongly encouraged to migrate to a supported PHP version even if we don't force you to do so.